Save game improvements
World files are like 10 times smaller now, and basically everything doesnt rip apart your memory for storage. Got some new fixes done including the addition of manpower which will be important later in development. Play button doesnt work for the moment and theres a lot of offset on the map which im working on. Happy holidays!